Bluetooth: Modify logic to ignore HID notifications
This change reverts the previous logic and delays the
ignoring of indications for HID devices, after the processing
of service change indications.
This is needed for keeping the update count equal to the
number of apps registered, so that service change indications
can be properly confirmed and followed by remote service
discovery.
This reverts commit a85bd580013944b321d88c3c5515a015c9f67f72.
CRs-Fixed: 982471

Change-Id: Iabba84aac1af6847bb44fd2169bd0870d8d14a55