HF CLIENT: Change to disable sniff with specific Phones.
When device is configured as hf client and connected to
AG, if AG goes into sniff, there may be disconnection
of hf connection with AG.We get information of AG's
manufacturer and model information using AT+CGMI
and AT+CGMM commands. We should disable sniff if AG's
model is known to have issues with sniff.

CRs-Fixed: 1021809
Change-Id: Iedfb8cbec7f716aadc83d511127834043eccfd57