Bluetooth: Implement hardware error timer for command timeout case
Root cause: As part of command timeout, ssr cleanup is initiated
and a special buffer is sent from the filter to SOC for intentional
crashing and collection of logs. The time provided before the
killing of stack is not sufficient for the whole dump to be
collected.

Fix: A hardware error timer is created which gets started in the
command timeout case immediately after sending ssr cleanup to the
vendor. The timer will expire after due time to kill the process.

Change-Id: I60cf55c9d4ae4078ee05bca89359de60091b3ba6
CRs-Fixed: 1022533