Bluetooth: Increase profile connection timeout
This change increases the timeout after the gatt
discovery complete from 1 sec to 5 sec.
For profiles such as HOGP, the profile level gatt connections
after the discovery complete callbacks sometime take more time,
due to which the "no app timeout" disconnects the physical link.
This change will hold the gatt connection used for remote discovery,
for 5 seconds, after the discovery is complete, and thus provide
sufficient time for the profile level connections to kick in.

CRs-Fixed: 1050047
Change-Id: I8f3c3ad162e3dadeb0e3215bf9fe6a20e4d9c425